Change UM top margin to 3.5cm (see commit body)
 - On printing, and without page re-adjustment to the printing area, the
page number was always cut. The 2.5cm required by UM standards are the
distance from the top page limit to the first line of text. Given that
the current template has a header, this header was always to close from
the limit of the page, because the value in the template is set
regarding the Latex text area, i.e., the area define by \textheight and
\textwidth.

- The template is now compliant to the standard, as the distance between
the top limit and the first line of text, in this case the text header,
is 2.5cm.